pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chimera-ipg"
version = "0.1.0"
description = "Intelligent Protocol Gateway for Project CHIMERA"
requires-python = ">=3.10"
dependencies = [
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "PyYAML>=6.0",
    "Faker>=24.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*", "scenarios*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]

